Small refactoring, move implementations into separate packages.
SQL code is now in `lsst.dax.apdb.sql`, Cassandra is in `...cassandra`.
