+++ tests/test_apdbSqlSchema.py
@@ -29,7 +29,7 @@ from typing import Any
 import lsst.utils.tests
 import sqlalchemy
 from lsst.dax.apdb.apdbSchema import ApdbTables
-from lsst.dax.apdb.apdbSqlSchema import ApdbSqlSchema, ExtraTables
+from lsst.dax.apdb.sql.apdbSqlSchema import ApdbSqlSchema, ExtraTables
 from lsst.dax.apdb.tests import update_schema_yaml
 from sqlalchemy import create_engine
